requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
google-genai>=1.10.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
PyPDF2>=3.0.0
websockets>=12.0
python-multipart>=0.0.6
itsdangerous>=2.2.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.1
supabase>=2.0.0
